Removed the components from the BasePage directory to a separate 'components' directory, renamed the 'hooks' directory to 'router'. Started work on replacing 'alert' windows with custom modal windows.
