+++ src/components/ModalWindow/ModalWindow.css
@@ -1,0 +1,18 @@
+.wrap {
+    margin-left: 15%;
+    width: 53%;
+    height: 100%;
+    position: fixed;
+    left: 0; top: 0; right: 0; bottom: 0;
+}
+  
+.modal {
+    position: absolute;
+    left: 36%;
+    top: 30%;
+    display: block;
+    width: 300px;
+    height: 200px;
+    background: rgb(202, 205, 209);
+}
+
